Implement EventComparator class
